stis extraction progress
stis_corrtag will read in a corrtag and re-derive columns.
filtering out events with stis data, only filtering airglow currently.

+++ lightcurve/stis.py
@@ -17,6 +17,7 @@ from astropy.io import fits as fits
 
 from .utils import expand_refname, enlarge
 from .stis_cal import map_image
+from .cos import extract_index
 from .version import version as  __version__
 
 #-------------------------------------------------------------------------------
@@ -30,11 +31,27 @@ def extract(filename, **kwargs):
     """
 
     step = kwargs.get('step', 1)
+    wlim = kwargs.get('wlim', (2, 10000))
+    xlim = kwargs.get('xlim', (0, 2048))
+    ylim = kwargs.get('ylim', (0, 2048))
+    filter_airglow = kwargs.get('filter_airglow', True)
+
+    hdu = fits.open(filename)
+
+    meta = {'source': filename,
+            'hdus': {'A': hdu},
+            'stepsize': step,
+            'wlim': wlim,
+            'xlim': xlim,
+            'ylim': ylim}
+
+    print("Using wlim: {}".format(wlim))
+    print("Using xlim: {}".format(xlim))
+    print("Using ylim: {}".format(ylim))
+    print("      step: {}".format(step))
 
     SECOND_PER_MJD = 1.15741e-5
 
-    hdu = fits.open(filename)
-    hdu = hdu
 
     time = hdu[1].data['time']
     #time = np.array([round(val, 3) for val in hdu[1].data['time']]).astype(np.float64)
@@ -53,9 +70,20 @@ def extract(filename, **kwargs):
     else:
         truncate = False
 
-    gross = np.histogram(time, all_steps)[0]
+    ### need to fix sdqflags
+    index = extract_index(hdu,
+                          xlim[0], xlim[1],
+                          ylim[0], ylim[1],
+                          wlim[0], wlim[1],
+                          0,
+                          filter_airglow=filter_airglow)
 
-    gross = gross
+    print("#{} events".format(len(index)))
+
+    gross = np.histogram(hdu['events'].data['time'][index],
+                          all_steps,
+                          weights=hdu['events'].data['epsilon'][index])[0]
+
     flux = np.zeros(gross.shape)
     background = np.zeros(gross.shape)
     mjd = hdu[1].header['EXPSTART'] + np.array(all_steps[:-1]) * SECOND_PER_MJD
@@ -72,7 +100,6 @@ def extract(filename, **kwargs):
 
     data = [times, mjd, bins, gross, background, flux]
     columns = ('times', 'mjd' ,'bins', 'gross', 'background', 'flux')
-    meta = {}
 
     return data, columns, meta
 
@@ -88,18 +115,27 @@ def stis_corrtag(tagfile, clean=True):
 
     """
     print("Creating STIS corrtag for {}".format(tagfile))
-    x1d_filename = tagfile.replace('_tag.fits', '_x1d.fits')
+    if '_tag.fits' in tagfile:
+        x1d_filename = tagfile.replace('_tag.fits', '_x1d.fits')
+    elif '_corrtag.fits' in tagfile:
+        x1d_filename = tagfile.replace('_corrtag.fits', '_x1d.fits')
 
     with fits.open(tagfile, 'readonly') as hdu:
         n_events = len(hdu[1].data['TIME'])
 
         time_data = hdu[1].data['TIME']
 
-        #-- Y data is same in raw and corrected coordinates
-        rawx_data = hdu[1].data['DETAXIS1']
-        rawy_data = hdu[1].data['AXIS2']
-        xcorr_data = hdu[1].data['AXIS1']
-        ycorr_data = hdu[1].data['AXIS2']
+        if 'DETAXIS1' in hdu[1].data.names:
+            #-- Y data is same in raw and corrected coordinates
+            rawx_data = hdu[1].data['DETAXIS1']
+            rawy_data = hdu[1].data['AXIS2']
+            xcorr_data = hdu[1].data['AXIS1']
+            ycorr_data = hdu[1].data['AXIS2']
+        else:
+            rawx_data = hdu[1].data['RAWX']
+            rawy_data = hdu[1].data['RAWY']
+            xcorr_data = hdu[1].data['XCORR'].astype(np.int32)
+            ycorr_data = hdu[1].data['YCORR'].astype(np.int32)
 
         #-- copy over the primary
         header0 = hdu[0].header.copy()
@@ -110,16 +146,24 @@ def stis_corrtag(tagfile, clean=True):
     if not os.path.exists(x1d_filename):
         print("Could not find associated extracted spectrum {}".format(x1d_filename))
         wave_data = np.ones(n_events) * hdu[0].header['CENTRWV']
+
     elif header0['OPT_ELEM'].startswith('E'):
         #-- Put in average wavelength for now
         with fits.open(x1d_filename) as x1d:
-            wave_data = np.ones(n_events) * x1d[1].data['wavelength'].mean()
+            #--initialize everything to 0 wavelength
+            wave_data = np.zeros(n_events)
+
+            #for order in x1d[1].data:
+            #    spec_center = order['a2center'] - order['shifta2']
+            #    spec_height = order['extrsize']
+
     else:
         #-- Grab wavelengths from the x1d file
         int_pix = np.array(map(int, map(round, xcorr_data))).astype(np.int32)
         int_pix = np.where(int_pix < 0, 0, int_pix)
-        int_pix = np.where(int_pix > 1023, 1023, int_pix)
-        print(int_pix)
+        int_pix = np.where(int_pix > 2047, 2047, int_pix)
+        int_pix //= 2
+
         with fits.open(x1d_filename) as x1d:
             wave_data = x1d[1].data['wavelength'][0][int_pix]
             print(wave_data.mean())
@@ -157,7 +201,7 @@ def stis_corrtag(tagfile, clean=True):
     hdu_out[1].header.extend(header1, end=True)
 
     hdu_out.writeto(tagfile.replace('_tag.fits', '_corrtag.fits'), clobber=True)
-    if clean:
+    if clean and '_tag.fits' in tagfile:
         print("Removing input tagfile")
         os.remove(tagfile)
 
@@ -196,16 +240,21 @@ def epsilon(tagfile):
                 image = image_hdu[1].data
 
                 if not image.shape == (2048, 2048):
-                    x_factor = 2048 / image.shape[1]
-                    y_factor = 2048 / image.shape[0]
+                    x_factor = 2048 // image.shape[1]
+                    y_factor = 2048 // image.shape[0]
 
                     print('Enlarging by {},{}'.format(x_factor, y_factor))
                     image = enlarge(image, x_factor, y_factor)
 
                 #--indexing is 1 off
-                epsilon_out *= map_image(image,
-                                         hdu[1].data['AXIS1'] - 1,
-                                         hdu[1].data['AXIS2'] - 1)
+                if 'AXIS1' in hdu[1].data.names:
+                    epsilon_out *= map_image(image,
+                                             hdu[1].data['AXIS1'] - 1,
+                                             hdu[1].data['AXIS2'] - 1)
+                else:
+                    epsilon_out *= map_image(image,
+                                             hdu[1].data['XCORR'].astype(np.int32) - 1,
+                                             hdu[1].data['YCORR'].astype(np.int32) - 1)
 
     return epsilon_out
 
@@ -258,8 +307,58 @@ def dqinit(tagfile):
 
         #-- Map to the events
         #-- indexing is 1 off
-        dq_out = map_image(dq_im,
-                           hdu[1].data['AXIS1'] - 1,
-                           hdu[1].data['AXIS2'] - 1)
+        if 'AXIS1' in hdu[1].data.names:
+            dq_out = map_image(dq_im,
+                               hdu[1].data['AXIS1'] - 1,
+                               hdu[1].data['AXIS2'] - 1)
+        else:
+            dq_out = map_image(dq_im,
+                               hdu[1].data['XCORR'].astype(np.int32) - 1,
+                               hdu[1].data['YCORR'].astype(np.int32) - 1)
 
     return dq_out
+
+#-------------------------------------------------------------------------------
+
+def crazy_histogram2d(x, y, bins=(2048, 2048)):
+    """  Bin events to a 2d image.
+
+    This is faster than the numpy version.
+
+    http://stackoverflow.com/questions/8805601/efficiently-create-2d-histograms-from-large-datasets
+
+    Parameters
+    ----------
+    x : array-like
+        x-coordinates to bin
+    y : array-like
+        y-coordinates to bin
+    bins : tuple, optional
+        bins of output image
+
+    Returns
+    -------
+    grid, x_bins, y_bins : (np.ndarray, np.ndarray, np.ndarray)
+        binned image, x bins, and y bins
+    """
+
+    try:
+        nx, ny = bins
+    except TypeError:
+        nx = ny = bins
+
+    xmin, xmax = x.min(), x.max()
+    ymin, ymax = y.min(), y.max()
+    dx = (xmax - xmin) / (nx - 1.0)
+    dy = (ymax - ymin) / (ny - 1.0)
+
+    weights = np.ones(x.size)
+
+    # Basically, this is just doing what np.digitize does with one less copy
+    xyi = np.vstack((x,y)).T
+    xyi = np.floor(xyi, xyi).T
+
+    # Now, we'll exploit a sparse coo_matrix to build the 2D histogram...
+    grid = scipy.sparse.coo_matrix((weights, xyi), shape=(nx, ny)).toarray()
+
+    return grid, np.linspace(xmin, xmax, nx), np.linspace(ymin, ymax, ny)
